Fixed a bug in the hashes_org module.
Fixed a bug in the fullcontact module.

Fixed a bug in the search mixin. Resolves issue #178. (@markmuir)

Added the certificate_transparency module. (@buffaloverflow)

Updated the hibp modules to account for rate limiting changes. Resolves issue #216.

Fixed a bug in the recon-rpc script. (@supertramp7331)

Updated the ssl_san module to leverage the ssltools API. (@bdno86)

Fixed a bug in the whois_miner module.

Modified the output that occurs when adding ports to the database.

Updated the censysio module to handle pagination. (@skew)

Split the twitter module into the twitter_mentions and twitter mentioned modules, added rate limiting to the search_twitter_api method, and configured both modules to store results in the database.

+++ modules/recon/contacts-credentials/hibp_paste.py
@@ -1,4 +1,6 @@
 from recon.core.module import BaseModule
+import os
+import time
 import urllib
 
 class Module(BaseModule):
@@ -8,7 +10,8 @@ class Module(BaseModule):
         'author': 'Tim Tomes (@LaNMaSteR53)',
         'description': 'Leverages the haveibeenpwned.com API to determine if email addresses have been published to various paste sites. Adds compromised email addresses to the \'credentials\' table.',
         'comments': (
-            'Paste sites supported: Pastebin, Pastie, or Slexy',
+            'Paste sites supported: Pastebin, Pastie, Slexy, Ghostbin, QuickLeak, JustPaste, AdHocUrl, and OptOut.'
+            'The HIBP API is rate limited to 1 request per 1.5 seconds.',
         ),
         'query': 'SELECT DISTINCT email FROM contacts WHERE email IS NOT NULL',
         'options': (
@@ -25,6 +28,7 @@ class Module(BaseModule):
             'Ghostbin': 'https://ghostbin.com/paste/%s/raw',
             'QuickLeak': 'http://www.quickleak.ir/%s',
             'JustPaste': 'https://justpaste.it/%s',
+            'AdHocUrl': '%s',
             }
         # retrieve status
         base_url = 'https://haveibeenpwned.com/api/v2/%s/%s'
@@ -44,15 +48,22 @@ class Module(BaseModule):
                     if paste['Source'] in sites:
                         fileurl = sites[paste['Source']] % (paste['Id'])
                         download = self.options['download']
+                    elif self.options['download'] == True:
+                        self.alert('Download not available for %s pastes.' % (paste['Source']))
                     self.alert('%s => Paste found! Seen in a %s on %s (%s).' % (account, paste['Source'], paste['Date'], fileurl))
                     if download == True:
                         resp = self.request(fileurl)
                         if resp.status_code == 200:
-                            filepath = '%s/%s_%s_%s.txt' % (self.workspace, account, paste['Source'], paste['Id'])
-                            dl = open(filepath, 'w')
-                            dl.write(resp.text.encode(resp.encoding) if resp.encoding else resp.text)
-                            dl.close()
+                            filepath = '%s/%s.txt' % (self.workspace, _safe_file_name(fileurl))
+                            if not os.path.exists(filepath):
+                                dl = open(filepath, 'w')
+                                dl.write(resp.text.encode(resp.encoding) if resp.encoding else resp.text)
+                                dl.close()
                             self.verbose('Paste stored at \'%s\'.' % (filepath))
                         else:
-                            self.output('Paste could not be downloaded (%s).' % (fileurl))
+                            self.alert('Paste could not be downloaded (%s).' % (fileurl))
                 self.add_credentials(account)
+            time.sleep(1.6)
+
+def _safe_file_name(s):
+    return "".join(c for c in s if c.isalnum()).rstrip()
